Write tests to getPageFromUrl function
