+++ src/utils/getPageFromUrl.ts
@@ -1,4 +1,7 @@
+import { URL } from 'url';
+
 export default function getPageFromUrl(url: string) {
-  const searchParams = new URLSearchParams(url);
+  const { searchParams } = new URL(url);
+
   return Number(searchParams.get('page'));
 }
